Two changes:
1. A new event was added, MHoverEvent which is sent when the mouse cursor enters or leaves a widget. When the cursor enter the widget, MHoverEvent's "enter" field (bool) is set to true and "other" field (Widget*) is set to the widget that previously had the mouse cursor over it (or NULL). When the cursor leaves the widget, MHoverEvent's "enter" field is false and "other" is set to the widget that now has the mouse cursor. Since MHoverEvent is derived from MouseEvent, it also contains everything you would expect. For MHoverEvent, a new event id was made, MHOVER_EVENT.

To ease the usage of hovering, a new field "hover" (bool) was added to Widget. This new field is set to true when the mouse is over the cursor and false otherwise. Currently this value is used by the Button to produce a "glow" effect when the mouse is over it or when it is in a "pressed" state but the mouse is not over it and to check if the mouse is over the button when the user releases the mouse button.

2. focused_window was afraid of the outside world, so it went back to utk_events.cc. For this, a new function was added, invalidate_widget(Widget *w) which is called by the Widget's destructor. invalidate_widget checks the passed widget object against focused_event, mouse_press_widget, mouse_grab_widget and last_hover_widget. If it is equal with any of these, they're set to NULL, in order to avoid corny crashes.

Also the code was cleaned up a bit (but only a bit, so don't call the neighbours).

--HG--
branch : ubertk
extra : convert_revision : svn%3A419d067c-91e1-11de-9a8c-5f207592b8ba/ubertk%4079