New properties and functions for the Widget class to set minimum and maximum sizes. set_min_size, set_max_size and the get_min/max_width/height function sets. set_size checks for these bounds and clamps the requested values (so modifying size.x and size.y manually is not a nice thing to do). Currently this is only used by the buttons (see below) to set a minimum size to 100x6. Later it will be also used for layout purposes.
Button resizing and creation improved a little. The bug mentioned in the previous log wasn't in buttons but i forgot to remove a ", 30" in test.cc which caused buttons to be always 30 pixels in width, so i was looking for an inexistent bug.

--HG--
branch : ubertk
extra : convert_revision : svn%3A419d067c-91e1-11de-9a8c-5f207592b8ba/ubertk%4092